+++ climaf/dataloc.py
@@ -13,6 +13,9 @@ from climaf.period import init_period
 from climaf.netcdfbasics import fileHasVar
 from clogging import clogger,dedent
 from operator import itemgetter
+import ftplib as ftp  
+import getpass 
+import netrc  
 
 locs=[]
 
@@ -27,14 +30,54 @@ class dataloc():
           model (str,optional): model name
           simulation (str,optional): simulation name
           frequency (str,optional): frequency
-          organization (str): name of the organization type, among 
-           those handled by :py:func:`~climaf.dataloc.selectLocalFiles`
-          url (list of strings): list of URLS for the data root directories
+          organization (str): name of the organization type, among those handled by :py:func:`~climaf.dataloc.selectFiles`
+          url (list of strings): list of URLS for the data root directories, local or remote
 
         Each entry in the dictionary allows to store :
         
-         - a list of path or URLS, which are root paths for
-           finding some sets of datafiles which share a file organization scheme
+         - a list of path or URLS (local or remote), which are root paths for
+           finding some sets of datafiles which share a file organization scheme.
+
+           - For remote data:
+           
+             url is supposed to be in the format 'protocol:user@host:path', but
+             'protocol' and 'user' are optional. So, url can also be 'user@host:path'
+             or 'protocol:host:path' or 'host:path'. ftp is default protocol.
+
+             If 'user' is given:
+           
+             - if 'host' is in $HOME/.netrc file, CliMAF check if corresponding
+               'login is the same of 'user'. If it is, CliMAF get associated
+               password; otherwise it prompt the user for entering password;
+             - if 'host' is not present in $HOME/.netrc file, CliMAF prompt the
+               user for entering password.
+
+             If 'user' is not given:
+           
+             - if 'host' is in $HOME/.netrc file, CliMAF get corresponding 'login'
+               as 'user' and also get associated password;
+             - if 'host' is not present in $HOME/.netrc file, CliMAF prompt the
+               user for entering 'user' and 'password'.
+
+             Remark: The .netrc file contains login and password used by the
+             auto-login process. It generally resides in the user's home directory
+             ($HOME/.netrc). So, it is highly recommended to supply this information
+             in .netrc file not to have to enter password in every request.
+
+             Warning: python netrc module does not handle multiple entries for a
+             single host. So, if netrc file has two entries for the same host, the
+             netrc module only returns the last entry.
+
+             We define two kinds of host: hosts with incremental files, e.g.
+             'beaufix'; and the others.
+
+             If one or all files returned by function :py:meth:`~climaf.classes.cdataset.listfiles`
+             are found in cache:
+
+             - in case of hosts with incremental files, files are only transferred
+               if their date on server is more recent than that found in cache;
+             - for other hosts, files found in cache are used
+
          - the name for the corresponding data files organization scheme. The current set of known
            schemes is :
 
@@ -50,8 +93,13 @@ class dataloc():
            - NetCDF model outputs as available during an ECLIS or ligIGCM simulation
            - ESGF
            
-         - the set of attribute values which which simulation's data are 
+         - the set of attribute values which simulation's data are 
            stored at that URLS and with that organization
+           
+           For remote files, filename pattern should include ${varname}, which is instanciated
+           by variable name or filenameVar (given via :py:func:`~climaf.classes.calias()`),
+           because files are not transferred to check if they contain the variable
+           
 
         For the sake of brievity, each attribute can have the '*'
         wildcard value; when using the dictionary, the most specific
@@ -66,6 +114,14 @@ class dataloc():
          - and declaring an exception for one simulation (here, both location and organization are supposed to be different)::
             
             >>> dataloc(project='PRE_CMIP6', model='IPSLCM-Z-HR', simulation='my_exp', organization='EM', url=['~/tmp/my_exp_data'])
+
+         - and declaring a project to access remote data (on multiple servers)::
+
+            >>> cproject('MY_REMOTE_DATA', ('frequency', 'monthly'), separator='|')
+            >>> dataloc(project='MY_REMOTE_DATA', organization='generic',url=['beaufix:/home/gmgec/mrgu/vignonl/L/${simulation}SFXYYYY.nc',
+            ... 'ftp:vignonl@hendrix:/home/vignonl/${model}/${variable}_1m_YYYYMM_YYYYMM_${model}.nc']),
+            >>> calias('MY_REMOTE_DATA','tas','tas',filenameVar='2T')
+            >>> tas=ds(project='MY_REMOTE_DATA', simulation='AMIPV6ALB2G', variable='tas', frequency='monthly', period='198101')        
 
          Please refer to the :ref:`example section <examples>` of the documentation for an example with each organization scheme
 
@@ -85,7 +141,8 @@ class dataloc():
         self.urls = map(os.path.expanduser,self.urls)
         alt=[]
         for u in self.urls :
-            if u[0] != '$' : alt.append(os.path.abspath(u))
+            #if u[0] != '$' : alt.append(os.path.abspath(u)) #lv
+            if u[0] != '$' and ':' not in u: alt.append(os.path.abspath(u))             
             else : alt.append(u)
         self.urls=alt
         # Register new dataloc only if not already registered
@@ -147,10 +204,10 @@ def isLocal(project, model, simulation, frequency) :
             if re.findall(".*:.*",l) : rep=False
     return rep
 
-def selectLocalFiles(**kwargs):
+def selectFiles(**kwargs):
     """
-    Returns the shortest list of (local) files which include the data
-    for the list of (facet,value) pairs provided
+    Returns the shortest list of (local or remote) files which include
+    the data for the list of (facet,value) pairs provided
 
     Method : 
     
@@ -218,8 +275,6 @@ def selectLocalFiles(**kwargs):
     # Assemble filenames in one single string
     return(string.join(rep))
 
-# u="/home/stephane/Bureau/climaf/examples/data/${simulation}/L/${simulation}SFXYYYY.nc"
-# selectGenericFiles(simulation="AMIPV6ALBG2", variable="tas", period="1980", urls=[u])
 
 def selectGenericFiles(urls, **kwargs):
     """
@@ -232,9 +287,12 @@ def selectGenericFiles(urls, **kwargs):
      - contain the ``variable`` provided in kwargs
 
      - match the `period`` provided in kwargs
-
-    In the pattern strings, no keyword is mandatory
-
+    
+    In the pattern strings, no keyword is mandatory. However, for remote files,
+    filename pattern should include ${varname}, which is instanciated by variable
+    name or ``filenameVar`` (given via :py:func:`~climaf.classes.calias()`), because
+    files are not transferred to check if they contain the variable
+    
     Example :
 
     >>> selectGenericFiles(project='my_projet',model='my_model', simulation='lastexp', variable='tas', period='1980', urls=['~/DATA/${project}/${model}/*${variable}*YYYY*.nc)']
@@ -270,24 +328,39 @@ def selectGenericFiles(urls, **kwargs):
     #
     for l in urls :
         # Instantiate keywords in pattern with attributes values
-        template=Template(l).safe_substitute(**kwargs)
+        if re.findall(".*:.*",l) : # remote data
+            template=Template(l.split(":")[-1]).safe_substitute(**kwargs)
+        else: # local data
+            template=Template(l).safe_substitute(**kwargs)
         #print "template after attributes replace : "+template
         #
         # Construct a pattern for globbing dates
-        temp2=template ; 
+        temp2=template ;
         for k in lkeys : temp2=temp2.replace(k,dt[k])
-        lfiles=glob.glob(temp2)
-        clogger.debug("Globbing %d files for varname on %s : "%(len(lfiles),temp2))
+        if re.findall(".*:.*",l) : # check if data is remote
+            lfiles=sorted(glob_remote_data(l, temp2))
+            clogger.debug("Globbing %d files for varname on %s : "%(len(lfiles),':'.join(l.split(":")[0:-1])+':'+temp2))
+        else: # local data
+            lfiles=sorted(glob.glob(temp2))
+            clogger.debug("Globbing %d files for varname on %s : "%(len(lfiles),temp2))
         #
         # If unsuccessful using varname, try with filenameVar
         if len(lfiles)==0 and "filenameVar" in kwargs and kwargs['filenameVar'] :
             kwargs['variable']=kwargs['filenameVar']
-            template=Template(l).safe_substitute(**kwargs)
+            
+            if re.findall(".*:.*",l) : # remote data
+                template=Template(l.split(":")[-1]).safe_substitute(**kwargs)
+            else: # local data
+                template=Template(l).safe_substitute(**kwargs)    
             temp2=template
             for k in lkeys : temp2=temp2.replace(k,dt[k])
             #
-            lfiles=glob.glob(temp2)
-            clogger.debug("Globbing %d files for filenamevar on %s: "%(len(lfiles),temp2))
+            if re.findall(".*:.*",l) : # check if data is remote
+                lfiles=sorted(glob_remote_data(l, temp2))
+                clogger.debug("Globbing %d files for filenamevar on %s: "%(len(lfiles),':'.join(l.split(":")[0:-1])+':'+temp2))
+            else: # local data
+                lfiles=sorted(glob.glob(temp2))
+                clogger.debug("Globbing %d files for filenamevar on %s: "%(len(lfiles),temp2))
 
         # Construct regexp for extracting dates from filename
         regexp=None
@@ -328,39 +401,106 @@ def selectGenericFiles(urls, **kwargs):
             else :
                 if ( 'frequency' in kwargs and ((kwargs['frequency']=="fx") or \
                     kwargs['frequency']=="seasonnal" or kwargs['frequency']=="annual_cycle" )) :
-                    if (l.find("${variable}")>=0) or variable=='*' or \
-                       fileHasVar(f,variable) or (variable != altvar and fileHasVar(f,altvar)) : 
+                    # local data
+                    if not re.findall(".*:.*",l) and \
+                       ( (l.find("${variable}")>=0) or variable=='*' or \
+                         fileHasVar(f,variable) or (variable != altvar and fileHasVar(f,altvar)) ) :
                         clogger.debug("adding fixed field :"+f)
                         rep.append(f)
+                    # remote data
+                    elif re.findall(".*:.*",l) :
+                        if (l.split(":")[-1].find("${variable}")>=0) or variable=='*' or \
+                           (variable != altvar and (f.find(altvar)>=0) ):
+                            clogger.debug("adding fixed field :"+':'.join(l.split(":")[0:-1])+':'+f)
+                            rep.append(':'.join(l.split(":")[0:-1])+':'+f)
+                        else:
+                            raise Climaf_Data_Error("For remote files, filename pattern (%s) should include ${varname} (which is instanciated by variable name or filenameVar)"%(':'.join(l.split(":")[0:-1])+':'+f))
                 else :
-                    clogger.warning("Cannot yet filter files re. time using only file content. TBD")
+                    clogger.info("Cannot yet filter files re. time using only file content.")
                     rep.append(f)
+            
             if (fperiod and period.intersects(fperiod)) or not regexp :
                 clogger.debug('Period is OK - Considering variable filtering on %s and %s for %s'%(variable,altvar,f)) 
                 # Filter against variable 
                 if (l.find("${variable}")>=0):
                     clogger.debug('appending %s based on variable in filename'%f)
                     rep.append(f)
-                    continue
-                if (f not in rep) and \
-                   (variable=='*' or "," in variable or fileHasVar(f,variable) or \
-                    (altvar != variable and fileHasVar(f,altvar))) :
-                    # Should check time period in the file if not regexp
-                    clogger.debug('appending %s based on multi-var or var exists in file '%f)
-                    rep.append(f)
-                    continue
+                    continue    
+                if (f not in rep):
+                    # local data
+                    if not re.findall(".*:.*",l) and \
+                        (variable=='*' or "," in variable or fileHasVar(f,variable) or \
+                        (altvar != variable and fileHasVar(f,altvar))) :
+                        # Should check time period in the file if not regexp
+                        clogger.debug('appending %s based on multi-var or var exists in file '%f)
+                        rep.append(f)
+                        continue
+                    # remote data
+                    elif re.findall(".*:.*",l) : 
+                        if variable=='*' or "," in variable or \
+                            (variable != altvar and (f.find(altvar)>=0) ):
+                            # Should check time period in the file if not regexp
+                            clogger.debug('appending %s based on multi-var or var exists in file '%(':'.join(l.split(":")[0:-1])+':'+f))
+                            rep.append(':'.join(l.split(":")[0:-1])+':'+f)
+                            continue
+                        else:
+                            raise Climaf_Data_Error("For remote files, filename pattern (%s) should include ${varname} (which is instanciated by variable name or filenameVar)"%(':'.join(l.split(":")[0:-1])+':'+f))
             else:
                 if not fperiod :
                     clogger.debug('not appending %s because period is None '%f)
                 else:
                     if not period.intersects(fperiod) :
                         clogger.debug('not appending %s because period doesn t intersect %s'%(f,period))
-
+                        
     return rep
 
 
+def glob_remote_data(url, pattern) :
+    """
+    Returns a list of path names that match pattern, for remote data
+    located at url
+    """
+    
+    if len(url.split(":")) == 3:
+        k=1
+    else:
+        k=0
+
+    if re.findall("@",url.split(":")[k]):
+        username=url.split(":")[k].split("@")[0]
+        host=url.split(":")[k].split("@")[-1]
+    else:
+        username=''
+        host=url.split(":")[k]
+
+    secrets = netrc.netrc()
+
+    if username:
+        if host in secrets.hosts:
+            login, account, password = secrets.authenticators( host )
+            if login != username: password = getpass.getpass("Password for host '%s' and user '%s': "%(host,username))
+        else:
+            password = getpass.getpass("Password for host '%s' and user '%s': "%(host,username))            
+    else:
+        if host in secrets.hosts:
+            username, account, password = secrets.authenticators( host )
+        else:
+            username = raw_input("Enter login for host '%s': " %host)
+            password = getpass.getpass("Password for host '%s' and user '%s': "%(host,username))
+
+    try : 
+        connect=ftp.FTP(host,username,password)
+        listfiles=connect.nlst(pattern)
+        connect.quit()
+        return(listfiles)
+    except ftp.all_errors as err_ftp:
+        print err_ftp
+        raise Climaf_Data_Error("Access problem for data %s on host '%s' and user '%s'" %(url,host,username))
+
+
+    
 def selectEmFiles(**kwargs) :
-    #POur A et L : mon, day1, day2, 6hLev, 6hPlev, 3h
+    #Pour A et L : mon, day1, day2, 6hLev, 6hPlev, 3h
     simulation=kwargs['simulation']
     frequency=kwargs['frequency']
     variable=kwargs['variable']