+++ scripts/regrid.sh
@@ -14,8 +14,8 @@ fieldout=$3
 option=${4:-remapbil}
 if [ -f $fieldgrid ] ; then 
     cdo griddes $fieldgrid > climaf_tmp_grid
-    cdo $option,climaf_tmp_grid $fieldin $fieldout 
-    rm climaf_tmp_grid
+    cdo $option,climaf_tmp_grid_$$ $fieldin $fieldout 
+    rm climaf_tmp_grid_$$
 else
     cdo $option,$fieldgrid $fieldin $fieldout 
 fi